+++ addons/crm/tests/__init__.py
@@ -3,6 +3,7 @@
 from . import test_crm_lead
 from . import test_crm_lead_notification
 from . import test_crm_lead_convert
+from . import test_crm_lead_convert_mass
 from . import test_crm_lead_lost
 from . import test_crm_lead_merge
 from . import test_crm_activity
